Rename in1 and in2 to indent_size and indent_no_tabs
