+++ example_files/example_lint_cfg.json
@@ -6,8 +6,8 @@
     "nsp_unary": {},
     "nsp_trailing": {},
     "long_lines": { "max_length": 80 },
-    "in1": { "indentation_spaces": 4 },
-    "in2": {},
+    "indent_size": { "indentation_spaces": 4 },
+    "indent_no_tabs": {},
     "in3": {},
     "in4": {},
     "in5": {},
