+++ src/lint/rules/tests/indentation/mod.rs
@@ -1,4 +1,4 @@
-mod in2;
+mod no_tabs;
 mod in3;
 mod in4;
 mod in5;
